pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "process-lab"
version = "0.1.0"
description = "A toolkit for optimizing New Relic Infrastructure Agent ProcessSample configurations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Proprietary"}
authors = [
    {name = "New Relic, Inc.", email = "info@newrelic.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.3.5",
    "requests>=2.30.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "pandas>=2.0.0",
    "jsonschema>=4.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "pytest-snapshot>=0.9.0",
    "pytest-benchmark>=4.0.0",
    "pytest-docker>=1.0.1",
    "ruff>=0.1.5",
    "mypy>=1.5.1",
    "bandit>=1.7.5",
    "hypothesis>=6.82.0",
    "cosmic-ray>=8.3.6",
    "tox>=4.11.3",
]

[project.scripts]
process-lab = "process_lab.cli:app"

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "B", "I", "W", "C90"]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "compose: marks tests that require docker-compose",
    "slow: marks tests that are slow to run",
]
